Change the project organization and code gen
Configure CubeMX to generate code for each peripheral in a separate
file (in which the handle can by accessed by including the .h file).

This new project structure will help to decouple the initialization of
peripheral from the main.c file. Thus helping the scheduler to
initialize its own hardware.